stream_readers refactor draft
